// src/app/sharedFeatures/enum/permission-enum.ts
export enum PermissionEnum {
 
    RoleList = 1,
    RoleView = 2,
    RoleAdd = 3,
    RoleEdit = 4,
    RoleDelete = 5,
    ChangeRolePermisstions = 14,
    GroupList = 6,
    AddGroup = 7,
    EditGroup = 8,
    UserList = 9,
    UserView = 10,
    UserAdd = 11,
    UserEdit = 12,
    UserDelete = 13,
    UserLoginList = 56,
    ChangeUserRoles = 57,
    SupplierList = 15,
    SupplierAdd = 16,
    SupplierEdit = 17,
    SupplierDelete = 18,
    ItemTypesList = 19,
    ItemTypesAdd = 20,
    ItemTypesEdit = 21,
    ItemTypesDelete = 22,
    ItemCategoriesList = 23,
    ItemCategoriesAdd = 24,
    ItemCategoriesEdit = 25,
    ItemCategoriesDelete = 26,
    ExchangeUnitsList = 27,
    ExchangeUnitsAdd = 28,
    ExchangeUnitsEdit = 29,
    ExchangeUnitsDelete = 30,
    ClientsList = 31,
    ClientsAdd = 32,
    ClientsEdit = 33,
    ClientsDelete = 34,
    RefrigeratorsList = 35,
    RefrigeratorsAdd = 36,
    RefrigeratorsEdit = 37,
    RefrigeratorsDelete = 38,
    AdditionDocumentList = 39,
    AdditionDocumentAdd = 40,
    AdditionDocumentEdit = 41,
    AdditionDocumentApprove = 42,
    WithdrawDocumentList = 43,
    WithdrawDocumentAdd = 44,
    WithdrawDocumentEdit = 45,
    WithdrawDocumentApprove = 46,
    InventoryDocumentList = 47,
    InventoryDocumentAdd = 48,
    InventoryDocumentView = 49,
    ManeuveringDocumentList = 47,
    ManeuveringDocumentAdd = 48,
    ManeuveringDocumentDelete = 49,
    SampleAnalysisDocumentList = 50,
    SampleAnalysisDocumentAdd = 51,
    SampleAnalysisDocumentDelete = 52,
    CleansingDocumentList = 53,
    CleansingDocumentAdd = 54,
    CleansingDocumentDelete = 55,
    CleansingDocumentApprove = 58,
    ManeuveringDocumentApprove = 59,
    SampleAnalysisDocumentApprove = 60,
    WithdrawDocumentDelete=61,
    Dashboard=100


}